pyproject.toml
[project]
name = "bookstore-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crispy-bootstrap5>=2024.10",
    "django>=5.1.6",
    "django-crispy-forms>=2.3",
    "django-debug-toolbar>=5.0.1",
    "django-extensions>=3.2.3",
]
